feat: splitted env conf for zod and edge runtimes
